\name{yeast}
\alias{yeast}
\docType{data}

\title{Example dataset for learning Gaussian Graphical Models by the IRO Algorithm} 

\description{
Genomic expression patterns in the yeast Saccharomyces cerevisiae responding to diverse environmental changes. The whole dataset consists of 173 samples collected under different environmental settings, and is available at http://www-genome.stanford.edu/yeast_stress/. It should be used in \code{YeastIRO(data,...)}.
}
\usage{data(yeast)}
\format{
\describe{
  \item{ yeast }{ a \emph{n}x\emph{p} Yeast Cell expression data.}
}}


\references{
Gasch, A.P., Spellman, P.T., Kao, C.M., Carmel-Harel, O., Eisen, M.B., Storz, G., Botstein, D., and Brown, P.O. (2000). Genomic expression programs in the response of yeast cells to environmental changes. Molecular Biology of the Cell, 11, 4241-4257.
}


\keyword{datasets}
